refactor: added alias path
