+++ tenant-registration/tsconfig-paths-bootstrap.js
@@ -1,0 +1,8 @@
+import { compilerOptions } from './tsconfig.json';
+import { register } from 'tsconfig-paths';
+
+const baseUrl = './dist';
+register({
+  baseUrl,
+  paths: compilerOptions.paths,
+});
